Adds API endpoints
Adds methods to `client.rb` for the four major endpoints in the API: content, tags, sections, and single item. Also moves the BASE_API_URL to `connection.rb` and `merge`s the the `api_key` into `options` in `request.rb`.

I now have a `Hashie::Mash` of data coming back when I make requests. Just trying to decide whether the method definitions are good as is or should accept arguments differently. For example, is it OK to require the user to run `@client.search(q: 'search string goes here')`, or should I add a separate argument that takes the search string that defaults to an empty string?

One other thing I've noticed is that you can pass the API garbage parameters and it will still make a valid request. For example, `@client.search(garbage: 'junk')` will respond as if `@client.search` was invoked rather than throwing back an invalid parameters error.

I'm not sure the best way to handle this, yet. Should I restrict parameters to a known, valid list and throw an error otherwise on my end, or should I just let it be? I think the former makes more sense, but I'll have to research what other developers do faced with a similar situation.